backend/agent_01_content_strategist.py
import os
import redis
import json
from supabase import create_client, Client
from qdrant_client import QdrantClient
import google.generativeai as genai
import time

# Environment variables
SUPABASE_URL = os.environ.get("SUPABASE_URL", "http://localhost:8000")
SUPABASE_KEY = os.environ.get("SUPABASE_KEY", "dummy_key")
REDIS_URL = os.environ.get("REDIS_URL", "redis://localhost:6379")
QDRANT_URL = os.environ.get("QDRANT_URL", "http://localhost:6333")
GEMINI_API_KEY = os.environ.get("GEMINI_API_KEY", "")

# Initialize clients
supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
redis_client = redis.Redis.from_url(REDIS_URL)
# qdrant = QdrantClient(url=QDRANT_URL) # Assuming we need to store/retrieve vector embeddings
if GEMINI_API_KEY:
    genai.configure(api_key=GEMINI_API_KEY)
# llm_model = genai.GenerativeModel('gemini-2.5-flash')

def scan_trends():
    """
    Mock function to simulate fetching data from Google Trends, Reddit, Ahrefs.
    """
    print("Agent 01: Scanning trends...")
    return [
        {"topic": "AI tools for marketing", "trend_score": 85}
    ]

def deduplicate_trends(trends):
    """
    Mock function using Qdrant for deduplication.
    """
    print("Agent 01: Deduplicating trends using Qdrant...")
    return trends  # In reality, this checks vector space for similar past briefs

def generate_brief(topic_data):
    """
    Use LLM to generate a content brief.
    """
    print(f"Agent 01: Generating brief for {topic_data['topic']}...")
    # Mocking LLM Output
    return {
        "title": f"Best {topic_data['topic']} in 2026",
        "target_keyword": topic_data['topic'].lower(),
        "secondary_keywords": ["guide", "tutorial", "best practices"],
        "outline": "H2 Introduction\nH2 Top Tools\nH2 Benefits\nH2 Use Cases\nH2 Conclusion",
        "angle": "Comprehensive review and ranking",
        "word_count_target": 1500,
        "trend_score": topic_data['trend_score']
    }

def run_agent_01():
    print("Starting Agent 01 - Content Strategist")
    raw_trends = scan_trends()
    unique_trends = deduplicate_trends(raw_trends)
    
    for trend in unique_trends:
        brief = generate_brief(trend)
        
        # Insert to DB
        response = supabase.table("content_briefs").insert({
            "title": brief["title"],
            "target_keyword": brief["target_keyword"],
            "secondary_keywords": brief["secondary_keywords"],
            "outline": brief["outline"],
            "angle": brief["angle"],
            "status": "PENDING",
            "word_count_target": brief["word_count_target"],
            "trend_score": brief["trend_score"]
        }).execute()
        
        # Publish event
        brief_id = response.data[0]['id']
        event_data = {
            "event": "content_briefs_ready",
            "brief_id": brief_id
        }
        redis_client.publish("content_events", json.dumps(event_data))
        print(f"Agent 01: Brief saved & event 'content_briefs_ready' published for brief ID {brief_id}")

if __name__ == "__main__":
    # Run every 4 hours, or just once for demo
    run_agent_01()
