backend/requirements.txt
supabase
redis
qdrant-client
google-generativeai
fastapi
uvicorn
requests
python-dotenv
